Java integration

+++ src/backend/TraceGenerator.ts
@@ -1,33 +1,31 @@
-import stringify from 'stringify-json';
-import { ExtensionContext, Uri, debug } from 'vscode';
-import * as ErrorMessages from '../ErrorMessages';
-import { Variables } from '../constants';
-import { getConfigValue, setContextState } from '../utils';
-import { getDebugConfigurationFor, registerDebugAdapterTracker } from './DebugAdapterTracker';
-import * as FileHandler from './FileHandler';
+import { Uri, commands, debug } from 'vscode';
 import Completer from '../Completer';
+import * as ErrorMessages from '../ErrorMessages';
+import * as FileHandler from '../FileHandler';
+import { getDebugConfigurationFor, registerDebugAdapterTracker } from './DebugAdapterTracker';
+
+export const linesWithClass: Array<number> = [];
 
 export class TraceGenerator {
     backendTrace: BackendTrace = [];
     file: Uri;
     language: SupportedLanguages;
     private fileContent: string;
-    private context: ExtensionContext;
-    private hash: string;
     private inTestingState: boolean;
 
-    constructor(file: Uri, fileContent: string, context: ExtensionContext, hash: string, inTestingState: boolean, language: SupportedLanguages) {
+    constructor(file: Uri, fileContent: string, inTestingState: boolean, language: SupportedLanguages) {
         this.file = file;
         this.fileContent = fileContent;
-        this.context = context;
-        this.hash = hash;
         this.language = language;
         this.inTestingState = inTestingState;
     }
 
     async generateTrace(): Promise<BackendTrace | undefined> {
         // PRE QUERIES
-        const tempFile = await FileHandler.duplicateFileAndExtendWithPass(this.file, this.fileContent);
+        if (this.language === 'python') {
+            setLinesWithClass(this.fileContent);
+        }
+        const tempFile = this.language === 'python' ? await FileHandler.duplicateFileAndExtendWithPass(this.file, this.fileContent) : this.file;
         if (!tempFile) {
             await ErrorMessages.showSpecificErrorMessage(ErrorMessages.ERR_TMP_FILE, this.inTestingState);
             return;
@@ -49,23 +47,47 @@ export class TraceGenerator {
 
         // FINISHING
         debugAdapterTracker.dispose();
-        await FileHandler.deleteFile(tempFile);
-        if (getConfigValue<boolean>('outputBackendTrace')) {
-            await FileHandler.createBackendTraceOutput(this.backendTrace, this.file);
+        if (this.language === 'python') {
+            await FileHandler.deleteFile(tempFile);
+            await commands.executeCommand('workbench.action.closeActiveEditor');
         }
-        await setContextState(
-            this.context,
-            Variables.HASH_KEY + this.file.fsPath,
-            this.hash
-        );
-        await setContextState(
-            this.context,
-            Variables.TRACE_KEY + this.file.fsPath,
-            stringify(this.backendTrace)
-        );
 
-        return this.backendTrace;
+        return this.language === 'java' ? filterUnnecessarySteps(this.backendTrace) : this.backendTrace;
     }
+}
+
+function setLinesWithClass(fileContent: string) {
+    fileContent.split("\n").forEach((line, index) => {
+        if (line.startsWith("class ")) {
+            linesWithClass.push(index + 1);
+        };
+    });
+}
+
+function filterUnnecessarySteps(trace: BackendTrace) {
+    return trace.filter(duplicates).filter(notSameLine);
+}
+
+function duplicates(outerTraceElement: BackendTraceElem, outerIndex: number, backendTrace: BackendTrace): boolean {
+    return !backendTrace.filter((innerTraceElement, innerIndex) => JSON.stringify(outerTraceElement, replacer) === JSON.stringify(innerTraceElement, replacer) && innerIndex < outerIndex).length;
+}
+
+function replacer(key: any, value: any) {
+    if (value instanceof Map) {
+        return {
+            dataType: 'Map',
+            value: Array.from(value.entries()),
+        };
+    } else {
+        return value;
+    }
+}
+
+function notSameLine(traceElement: BackendTraceElem, index: number, backendTrace: BackendTrace): boolean {
+    if (index === 0) {
+        return true;
+    }
+    return traceElement.line !== backendTrace[index - 1].line;
 }
 
 async function initializeAdapterForActiveDebugSession(language: SupportedLanguages): Promise<Capabilities> {
